requirements.txt
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
streamlit
fastapi
uvicorn
python-dotenv
pydantic
uuid
python-multipart